backend/requirements.txt
Django==3.2.3
django-filter==23.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==10.1.0
python-dotenv==1.0.1
PyYAML==6.0
PyJWT==2.8.0
reportlab==4.0.7
requests==2.31.0
tqdm==4.66.1
